refactor(script): promeut script_automatheque hors de util (#41)
`script_automatheque` (+ alias `script`) et `ScriptAutomatheque` sont l'API
phare du projet (« se faciliter le scripting ») : leur place n'est pas parmi
les utilitaires bas niveau de `util` (fichier, reessaye, repertoire…).

- Nouveau module dédié `automatheque/script.py` (pas le __init__ du paquet,
  pour ne pas tirer docopt à chaque `import automatheque`).
- `automatheque/util/script.py` devient un shim : ré-exporte les mêmes objets
  et émet un DeprecationWarning à l'import (suppression à terme).
- Imports internes migrés vers le nouveau chemin (essai, factrice.app, tests).
- README paquet mis à jour (nouveau chemin + alias court `script`).
- Tests : import nouveau chemin, shim émet le DeprecationWarning, ré-exports
  pointant vers les mêmes objets.

Pré-1.0 : le shim assure la transition sans casser les scripts existants.
