+++ src/automatheque/CHANGELOG
@@ -9,6 +9,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+* `script_automatheque` / `script` / `ScriptAutomatheque` sont promus de
+  `automatheque.util.script` vers **`automatheque.script`** : c'est l'API phare
+  du projet, pas un utilitaire bas niveau. `automatheque.util.script` reste
+  importable (shim) mais émet un `DeprecationWarning` et sera supprimé
+  ultérieurement. Nouveau chemin : `from automatheque.script import script`.
+  — cf. #41
+
 ## [0.12.0] - 2026-07-15
 
 ### Added
